Relocate operators from smart playlist file to new file
